fixed the issue pointed by @marleeng

+++ server/src/main/java/org/linkeddatafragments/datasource/hdt/HdtRequestProcessor.java
@@ -48,9 +48,9 @@ public class HdtRequestProcessor extends AbstractRequestProcessor<RDFNode, Strin
    * HDT Datasource
    */
   private final HDT datasource;
-  private HashMap<Integer, HashMap<Integer, Integer>> predicateSubjectRanks;
-  private HashMap<Integer, Integer[]> predicateSubjectsSorted;
-  private HashMap<Integer, HashSet<Integer>> predicateSubjects;
+  private HashMap<Long, HashMap<Long, Integer>> predicateSubjectRanks;
+  private HashMap<Long, Long[]> predicateSubjectsSorted;
+  private HashMap<Long, HashSet<Long>> predicateSubjects;
   /**
    * The dictionary
    */
@@ -66,14 +66,14 @@ public class HdtRequestProcessor extends AbstractRequestProcessor<RDFNode, Strin
   public HdtRequestProcessor(String hdtFile) throws IOException, ClassNotFoundException {
     datasource = HDTManager.mapIndexedHDT(hdtFile, null); // listener=null
     dictionary = new NodeDictionary(datasource.getDictionary());
-    predicateSubjectRanks = new HashMap<Integer, HashMap<Integer, Integer>>();
-    predicateSubjectsSorted = new HashMap<Integer, Integer[]>();
+    predicateSubjectRanks = new HashMap<Long, HashMap<Long, Integer>>();
+    predicateSubjectsSorted = new HashMap<Long, Long[]>();
     initializeNumericPredicateIndexes(hdtFile);
 
-    predicateSubjects = new HashMap<Integer, HashSet<Integer>>();
-    for (int predicateId : predicateSubjectRanks.keySet()) {
+    predicateSubjects = new HashMap<Long, HashSet<Long>>();
+    for (long predicateId : predicateSubjectRanks.keySet()) {
       predicateSubjects.put(predicateId,
-          new HashSet<Integer>(predicateSubjectRanks.get(predicateId).keySet()));
+          new HashSet<Long>(predicateSubjectRanks.get(predicateId).keySet()));
     }
   }
 
@@ -82,52 +82,32 @@ public class HdtRequestProcessor extends AbstractRequestProcessor<RDFNode, Strin
       throws IOException, ClassNotFoundException {
     File psrObjectFile = new File(hdtFile + ".psr");
     File pssObjectFile = new File(hdtFile + ".pss");
-    /*
-     * Kryo kryo = new Kryo();
-     * 
-     * MapSerializer serializer = new MapSerializer(); kryo.register(HashMap.class, serializer);
-     * serializer.setKeyClass(Integer.class, kryo.getSerializer(Integer.class));
-     * serializer.setKeysCanBeNull(false); MapSerializer innerMapSerializer = new MapSerializer();
-     * serializer.setValueClass(HashMap.class, innerMapSerializer);
-     * innerMapSerializer.setKeyClass(Integer.class, kryo.getSerializer(Integer.class));
-     * kryo.register(BitSet.class, kryo.getSerializer(BitSet.class));
-     * 
-     * innerMapSerializer.setValueClass(SubjectBitSets.class, new Serializer<SubjectBitSets>() {
-     * public void write(Kryo kryo, Output output, SubjectBitSets object) { kryo.writeObject(output,
-     * object.getGreaterThanBitset()); kryo.writeObject(output, object.getEqualBitset()); }
-     * 
-     * public SubjectBitSets read(Kryo kryo, Input input, Class<SubjectBitSets> type) {
-     * SubjectBitSets subjectBitSets = new SubjectBitSets();
-     * subjectBitSets.setGreaterThanBitset(kryo.readObject(input, BitSet.class));
-     * subjectBitSets.setEqualBitset(kryo.readObject(input, BitSet.class)); return subjectBitSets; }
-     * });
-     */
 
     if (psrObjectFile.isFile()) {
       System.out.println("Psr file exists. Trying to initialize...");
       FSTObjectInput in = new FSTObjectInput(new FileInputStream(psrObjectFile));
-      predicateSubjectRanks = (HashMap<Integer, HashMap<Integer, Integer>>) in.readObject();
+      predicateSubjectRanks = (HashMap<Long, HashMap<Long, Integer>>) in.readObject();
       in.close(); // required !
       in = new FSTObjectInput(new FileInputStream(pssObjectFile));
-      predicateSubjectsSorted = (HashMap<Integer, Integer[]>) in.readObject();
+      predicateSubjectsSorted = (HashMap<Long, Long[]>) in.readObject();
       in.close(); // required !
 
       System.out.println(predicateSubjectRanks.size());
       System.out.println("Initialized numeric attribute index.");
     } else {
-      HashMap<Integer, HashMap<Integer, Float>> tripleIDValues =
-          new HashMap<Integer, HashMap<Integer, Float>>();
+      HashMap<Long, HashMap<Long, Float>> tripleIDValues =
+          new HashMap<Long, HashMap<Long, Float>>();
       Model emptyModel = ModelFactory.createDefaultModel();
-      HashSet<Integer> predicatesWithFloatValues = new HashSet<Integer>();
-      HashSet<Integer> predicatesWithStringValues = new HashSet<Integer>();
+      HashSet<Long> predicatesWithFloatValues = new HashSet<Long>();
+      HashSet<Long> predicatesWithStringValues = new HashSet<Long>();
       Iterator<TripleID> tripleIterator = datasource.getTriples().searchAll();
       while (tripleIterator.hasNext()) {
         TripleID tripleId = tripleIterator.next();
-        int predicateId = tripleId.getPredicate();
+        long predicateId = tripleId.getPredicate();
         if (predicatesWithStringValues.contains(predicateId)) {
           continue;
         }
-        int subjectId = tripleId.getSubject();
+        long subjectId = tripleId.getSubject();
 
         Triple triple = Utils.toTriple(dictionary, tripleId);
         Statement st = emptyModel.asStatement(triple);
@@ -158,7 +138,7 @@ public class HdtRequestProcessor extends AbstractRequestProcessor<RDFNode, Strin
           if (tripleIDValues.containsKey(predicateId)) {
             tripleIDValues.get(predicateId).put(subjectId, val);
           } else {
-            HashMap<Integer, Float> innerHashMap = new HashMap<Integer, Float>();
+            HashMap<Long, Float> innerHashMap = new HashMap<Long, Float>();
             innerHashMap.put(subjectId, val);
             tripleIDValues.put(tripleId.getPredicate(), innerHashMap);
           }
@@ -169,26 +149,26 @@ public class HdtRequestProcessor extends AbstractRequestProcessor<RDFNode, Strin
       System.out
           .println("Number of predicates with literal values: " + predicatesWithFloatValues.size());
       ArrayList<SubjectIdWithValue> tripleValues = new ArrayList<SubjectIdWithValue>();
-      for (int predicateId : predicatesWithFloatValues) {
+      for (long predicateId : predicatesWithFloatValues) {
         System.out.println("creating bitmaps for predicate " + predicateId);
         System.out.println(dictionary.getNode(predicateId, TripleComponentRole.PREDICATE).getURI());
-        Set<Integer> subjectIds = tripleIDValues.get(predicateId).keySet();
-        for (int subjectId : subjectIds) {
+        Set<Long> subjectIds = tripleIDValues.get(predicateId).keySet();
+        for (long subjectId : subjectIds) {
           tripleValues.add(
               new SubjectIdWithValue(subjectId, tripleIDValues.get(predicateId).get(subjectId)));
         }
         Collections.sort(tripleValues, new SubjectIdWithValueComparator(SkylinePrefFunc.MAX));
-        Integer[] sortingResult = new Integer[tripleValues.size()];
+        Long[] sortingResult = new Long[tripleValues.size()];
         for (int i = 0; i < sortingResult.length; i++) {
           sortingResult[i] = tripleValues.get(i).getSubjectId();
         }
         predicateSubjectsSorted.put(predicateId, sortingResult);
-        HashMap<Integer, Integer> subjectRanks = new HashMap<Integer, Integer>();
+        HashMap<Long, Integer> subjectRanks = new HashMap<Long, Integer>();
         int numOfTriples = tripleValues.size();
         int lastIdx = 0;
         float lastVal = tripleValues.get(0).getValue();
         for (int i = 0; i < numOfTriples; i++) {
-          int subjectId = tripleValues.get(i).getSubjectId();
+          long subjectId = tripleValues.get(i).getSubjectId();
           float currVal = tripleValues.get(i).getValue();
           if (lastVal == currVal) {
             subjectRanks.put(subjectId, lastIdx);
@@ -214,28 +194,6 @@ public class HdtRequestProcessor extends AbstractRequestProcessor<RDFNode, Strin
     }
   }
 
-  /**
-   * @param statements
-   * @return
-   * 
-   *         private HashMap<Integer, SubjectBitSets> createBitmapsForSubjects(
-   *         ArrayList<SubjectIdWithValue> subjectsWithValue) { HashMap<Integer, SubjectBitSets>
-   *         currentHashMap = new HashMap<Integer, SubjectBitSets>(); for (int i = 0; i <
-   *         subjectsWithValue.size(); i++) { SubjectIdWithValue currSubject =
-   *         subjectsWithValue.get(i); int currSubjectId = currSubject.getSubjectId();
-   *         HashSet<Integer> subjectIdsWithGreaterValues = new HashSet<Integer>(); HashSet<Integer>
-   *         subjectIdsWithEqualValues = new HashSet<Integer>(); boolean isDone = false; for (int j
-   *         = 0; j < subjectsWithValue.size() && !isDone; j++) { SubjectIdWithValue secondSubject =
-   *         subjectsWithValue.get(j); int secondSubjectId = secondSubject.getSubjectId(); if (i > j
-   *         && currSubject.getValue() != secondSubject.getValue()) {
-   *         subjectIdsWithGreaterValues.add(secondSubjectId); } else if (i > j) {
-   *         subjectIdsWithEqualValues.add(secondSubjectId); } else if (i < j &&
-   *         currSubject.getValue() == secondSubject.getValue()) {
-   *         subjectIdsWithEqualValues.add(secondSubjectId); } else if (i == j) { continue; } else {
-   *         isDone = true; } } currentHashMap.put(currSubjectId, new
-   *         SubjectBitSets(subjectIdsWithGreaterValues, subjectIdsWithEqualValues)); } return
-   *         currentHashMap; }
-   */
   /**
    *
    * @param request
